Made scrollDelta a float for high res scroll wheels

+++ src/ccore/core/common/interface/event.h
@@ -54,7 +54,7 @@ typedef struct {
 		ccMouseButtonType mouseButton;
 		ccPoint mousePosition;
 		unsigned int keyCode;
-		int scrollDelta;
+		float scrollDelta;
 	};
 } ccEvent;
